+++ CHANGELOG.md
@@ -38,6 +38,8 @@ The calls to these functions were modified accordingly.
 19/02/26
 - *processing.interactive_phase_1D*: modified zoom computation, added +90/-90 buttons
 - *processing.interactive_phase_2D*: modified zoom computation, added +90/-90 buttons
+20/02/26
+- *fit.get_region*: completely rewritten, now can return multiple regions
 
 ## Modified classes:
 12/02/26